refactor: Layoutをcomponentに移動
