+++ pages/account/[userId]/theme/index.tsx
@@ -1,7 +1,6 @@
+import { SettingLayout } from "@component/Layout/setting";
 import { Theme } from "@component/theme";
 import type { NextPage } from "next";
-
-import { SettingLayout } from "../../../layout/setting";
 
 const Index: NextPage = () => (
   <SettingLayout title={"テーマ"}>
